Initial Conversion to Monad Transformers.
This is the conversion to using monad transformers instead of the IORef
technique from the book. This was my first ever introduction to monad
transformers so it's a bit...interesting.

I've come a long way since this but I'm happy to share it with the world
and go from here.